Changed code to use HC-05 Bluetooth module
We were having too many issues using the BLE drivers and the BLE device built into the Artemis.  We switched to using the HC-05 Bluetooth module.  While this adds an extra component to the RSA, it eliminates the Artemis from the Handheld Device and allows for a much simpler wireless implementation.  With minor updates to the RSA motherboards design, it would be a seamless implementation.