Massive Overhaul
This is a complete re-write of the Motherboard code.  This version of the code should support multiple sensors at one time.  The ADIN1110s on the motherboard are initialized and data is gathered from them in a round-robin format.  Then data is sent to the handheld in a similar manner to limit the latency in data transmitted.